Playground React migration Phase 5: docs examples + polish, remove old main.ts
- data/docs.ts + docs-view.tsx: cli / lifo-pkg / build-pkg doc blocks (hideCode)
- registry: all 18 examples now real; comingSoon stub removed
- Code-split every example via React.lazy + a Suspense boundary in ExampleHost
  (Monaco/xterm/templates load only on first activation)
- vite: manualChunks split xterm + react-vendor into cacheable chunks
  (entry ~30kB vs the prior monolithic 1.24MB); Monaco left dynamic
- Delete the old 2725-line vanilla main.ts (fully superseded)

Migration complete: Vite + React 19 + shadcn/ui + Tailwind v4, resizable
columns and vertical splits everywhere, Tokyo Night look preserved.
