+++ apps/playground/src/examples/build-pkg.tsx
@@ -1,0 +1,6 @@
+import { DocsView } from '@/components/docs-view';
+import { DOC_BUILD_PKG } from '@/data/docs';
+
+export default function BuildPkgExample() {
+  return <DocsView title="Build Lifo Packages" html={DOC_BUILD_PKG} />;
+}
